Add local_access_cpu submodule for data module
