+++ hoomd/data/__init__.py
@@ -1,5 +1,6 @@
 from .local_access import (
-    LocalSnapshot, AngleLocalAccess, BondLocalAccess,
-    ConstraintLocalAccess, DihedralLocalAccess, ImproperLocalAccess,
-    PairLocalAccess, ParticleLocalAccess)
+    AngleLocalAccessBase, BondLocalAccessBase, ConstraintLocalAccessBase,
+    DihedralLocalAccessBase, ImproperLocalAccessBase, PairLocalAccessBase,
+    ParticleLocalAccessBase)
+from .local_access_cpu import LocalSnapshot
 from .local_access_gpu import LocalSnapshotGPU
